IE8 css and view modify

+++ src/main/webapp/WEB-INF/view/bss/cs/purchaseContract/draftlist.jsp
@@ -15,6 +15,8 @@
 	<meta http-equiv="description" content="This is my page">
   <script type="text/javascript">
   $(function(){
+	  var status = "${purCon.status}";
+	  $("#status").val(status);
 	  laypage({
 		    cont: $("#pagediv"), //容器。值支持id名、原生dom对象，jquery对象,
 		    pages: "${list.pages}", //总页数
@@ -85,14 +87,26 @@
     
   	function delDraft(){
     	var ids =[]; 
+    	var status = [];
+    	var flag = true;
 		$('input[name="chkItem"]:checked').each(function(){ 
 			ids.push($(this).val()); 
+			status.push($(this).parent().next().text());
 		}); 
 		if(ids.length>0){
-			layer.confirm('您确定要删除吗?', {title:'提示',offset: ['222px','360px'],shade:0.01}, function(index){
-				layer.close(index);
-				window.location.href="${pageContext.request.contextPath}/purchaseContract/deleteDraft.html?ids="+ids;
-			});
+			for(var i=0;i<status.length;i++){
+				if(status[i]==2){
+					flag = false;
+					layer.alert("正式合同不可删除",{offset: ['222px', '390px'], shade:0.01});
+				}
+			}
+			if(flag){
+				layer.confirm('您确定要删除吗?', {title:'提示',offset: ['222px','360px'],shade:0.01}, function(index){
+					layer.close(index);
+					window.location.href="${pageContext.request.contextPath}/purchaseContract/deleteDraft.html?ids="+ids;
+				});
+			}
+			
 		}else{
 			layer.alert("请选择要删除的信息",{offset: ['222px', '390px'], shade:0.01});
 		}
@@ -114,21 +128,28 @@
   		$("#purchaseDepName").val("");
   		$("#year").val("");
   		$("#budgetSubjectItem").val("");
+  		$("#status").val("");
   	}
   	
   	function updateDraft(){
   		var ids =[]; 
+  		var status = "";
 		$('input[name="chkItem"]:checked').each(function(){ 
 			ids.push($(this).val()); 
+			status = ($(this).parent().next().text());
 		}); 
 		if(ids.length>0){
 			if(ids.length>1){
 				layer.alert("只可选择一条修改",{offset: ['222px', '390px'], shade:0.01});
 			}else{
-				window.location.href="${pageContext.request.contextPath}/purchaseContract/createDraftContract.html?ids="+ids;
+				if(status!="2"){
+					window.location.href="${pageContext.request.contextPath}/purchaseContract/createDraftContract.html?ids="+ids;
+				}else{
+					layer.alert("只可修改暂存和草案的合同",{offset: ['222px', '390px'], shade:0.01});
+				}
 			}
 		}else{
-			layer.alert("请选择要修改的草稿",{offset: ['222px', '390px'], shade:0.01});
+			layer.alert("请选择要修改的合同",{offset: ['222px', '390px'], shade:0.01});
 		}
   	}
   	
@@ -163,55 +184,62 @@
 			layer.alert("请选择要生成的合同草稿",{offset: ['222px', '390px'], shade:0.01});
 		}
   	}
+  	var ind;
+  	function createDraftContract(){
+  		var ids =[];
+  		var status = "";
+  		flag = true;
+  		$('input[name="chkItem"]:checked').each(function(){ 
+			ids.push($(this).val()); 
+			status=($(this).parent().next().text());
+		}); 
+		if(ids.length>0){
+			if(ids.length>1){
+				layer.alert("只可选择一条暂存合同生成",{offset: ['222px', '390px'], shade:0.01});
+			}else{
+				if(status!=0){
+					flag = false;
+					layer.alert("只可选择暂存的合同生成",{offset: ['222px', '390px'], shade:0.01});
+				}
+				if(flag){
+					ind = layer.open({
+						shift: 1, //0-6的动画形式，-1不开启
+					    moveType: 1, //拖拽风格，0是默认，1是传统拖动
+					    title: ['生成草案所需信息','border-bottom:1px solid #e5e5e5'],
+					    shade:0.01, //遮罩透明度
+						type : 1,
+						skin : 'layui-layer-rim', //加上边框
+						area : [ '40%', '300px' ], //宽高
+						content : $('#numberWin'),
+						offset: ['10%', '25%']
+					});
+				}
+			}
+		}else{
+			layer.alert("请选择要生成的暂存合同",{offset: ['222px', '390px'], shade:0.01});
+		}
+  	}
   	function save(){
   		var ids =[];
   		$('input[name="chkItem"]:checked').each(function(){ 
 			ids.push($(this).val()); 
 		}); 
   		$("#ids").val(ids);
-		var apN = $("#apN").val();
-		var picFile = $("#fi").val();
-		var picFiles = picFile.split(".");
-		var pic = picFiles[picFiles.length-1];
-		var formalGitAt = $("#formalGitAt").val();
-		var formalReviewedAt = $("#formalReviewedAt").val();
-		var flag = false;
-		var news = "";
-		if(formalGitAt!=null && formalGitAt!=""){
-			flag = true;
-		}else{
-			flag = false;
-			news+="请填写上报时间";
-		}
-		if(formalReviewedAt!=null && formalReviewedAt!=""){
-			flag = true;
-		}else{
-			flag = false;
-			news+="请填写报批时间";
-		}
-		if(apN!=null && apN!=''){
-			flag = true;
-		}else{
-			flag = false;
-			news+="请填写合同批准文号,";
-		}
-		if(pic!=null && pic!=''){
-			if(pic=='bmp' || pic=='png' || pic=='gif' && pic=='jpg' && pic=='jpeg'){
-				flag=true;
-			}else{
-				flag=false;
-				news+="上传的附件类型不正确";
-			}
-		}else{
-			flag=false;
-			news+="请上传批准电子扫描件,";
-		}
-		
-		if(flag){
-			$("#contractForm").submit();
-		}else{
-			layer.alert(news,{offset: ['55%', '40%'], shade:0.01});
-		}
+  		$.ajax({
+  			url:"${pageContext.request.contextPath}/purchaseContract/toValidRoughContract.html",
+  			data:$('#form2').serialize(),
+  			type:"post",
+  			dataType:"json",
+  			success:function(data){
+  				if(data==1){
+					$("#form2").submit();
+				}else{
+					var obj = new Function("return" + data)();
+					$("#gitTime").text(obj.gitAt);
+					$("#reviewTime").text(obj.reviewAt);
+				}
+  			}
+  		});
 	}
 	
 	function cancel(){
@@ -227,6 +255,10 @@
 				time : 4000    //默认消息框不关闭
 		});//去掉msg图标
   	}
+	
+	function updateModel(){
+		window.location.href="${pageContext.request.contextPath}/templet/search.html?temType=合同模板";
+	}
   </script>
   </head>
   
@@ -259,6 +291,14 @@
 	      <li class="fl"><label class="fl">采购机构：</label><span><input type="text" value="${purCon.purchaseDepName }" id="purchaseDepName" name="purchaseDepName" class="mb0 mt5"/></span></li>
 	      <li class="fl"><label class="fl">年度：</label><span><input type="text" value="${purCon.year_string }" id="year" name="year_string" class="mb0 mt5 w100"/></span></li>
 	      <li class="fl"><label class="fl">项级预算科目：</label><span><input type="text" value="${purCon.budgetSubjectItem }" id="budgetSubjectItem" name="budgetSubjectItem" class="mb0 mt5 w100"/></span></li>
+	      <li class="fl"><label class="fl">状态：</label><span>
+	      	<select id="status" name="status" class="mb0 mt5 w100">
+	      		<option value="">--请选择--</option>
+	      		<option value="0">暂存</option>
+	      		<option value="1">草案</option>
+	      		<option value="2">正式</option>
+	      	</select>
+	      </span></li>
     	  <div class="fl col-md-12 tc mt10">
     	    <input type="submit" class="btn" value="查询"/>
     	    <input type="button" onclick="resetForm()" class="btn" value="重置"/>
@@ -271,6 +311,8 @@
    	  	  <button class="btn btn-windows edit" onclick="updateDraft()">修改</button>
    	  	  <button class="btn btn-windows delete" onclick="delDraft()">删除</button>
 	      <button class="btn" onclick="createContract()">生成正式合同</button>
+	      <button class="btn" onclick="createDraftContract()">生成草案合同</button>
+	      <button class="btn" onclick="updateModel()">更新合同模板</button>
 	      <div class="fr mt5 b">
 	      	项目总金额：${contractSum}
 	      </div>
@@ -280,88 +322,84 @@
 		<thead>
 			<tr>
 				<th class="info w30"><input id="checkAll" type="checkbox" onclick="selectAll()" /></th>
+				<th class="tnone"></th>
 			    <th class="info w50">序号</th>
 			    <th class="info">合同编号</th>
 				<th class="info">合同名称</th>
 				<th class="info">合同金额</th>
 				<th class="info">项目名称</th>
-				<th class="info">供应商名称</th>
-				<th class="info">采购机构</th>
-				<th class="info">需求部门</th>
 				<th class="info">计划文件号</th>
 				<th class="info">预算</th>
 				<th class="info">年度</th>
 				<th class="info">项级预算科目</th>
+				<th class="info">状态</th>
 			</tr>
 		</thead>
 		<c:forEach items="${draftConList}" var="draftCon" varStatus="vs">
 			<tr>
 				<td class="tc pointer"><input onclick="check()" type="checkbox" name="chkItem" value="${draftCon.id}" /></td>
+				<td class="tnone">${draftCon.status}</td>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${(vs.index+1)+(list.pageNum-1)*(list.pageSize)}</td>
 				<c:set value="${draftCon.code}" var="code"></c:set>
 				<c:set value="${fn:length(code)}" var="length"></c:set>
 				<c:if test="${length>7}">
-					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" onmouseover="titleMouseOver('${code}',this)" onmouseout="titleMouseOut();">${fn:substring(code,0,7)}...</td>
+					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" title="${code}">${fn:substring(code,0,7)}...</td>
 				</c:if>
 				<c:if test="${length<=7}">
-					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" onmouseover="titleMouseOver('${code}',this)" onmouseout="titleMouseOut();">${code}</td>
+					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" title="${code}">${code}</td>
 				</c:if>
 				<c:set value="${draftCon.name}" var="name"></c:set>
 				<c:set value="${fn:length(name)}" var="length"></c:set>
 				<c:if test="${length>9}">
-					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" onmouseover="titleMouseOver('${name}',this)" onmouseout="titleMouseOut();">${fn:substring(name,0,9)}...</td>
+					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" title="${name}">${fn:substring(name,0,9)}...</td>
 				</c:if>
 				<c:if test="${length<=9}">
-					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" onmouseover="titleMouseOver('${name}',this)" onmouseout="titleMouseOut();">${name}</td>
+					<td onclick="showDraftContract('${draftCon.id}')" class="pointer" title="${name}">${name}</td>
 				</c:if>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.money}</td>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.projectName}</td>
-				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.supplierDepName}</td>
-				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.purchaseDepName}</td>
-				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.demandSector}</td>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.documentNumber}</td>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.budget}</td>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.year}</td>
 				<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">${draftCon.budgetSubjectItem}</td>
+				<c:if test="${draftCon.status==0}">
+					<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">暂存</td>
+				</c:if>
+				<c:if test="${draftCon.status==1}">
+					<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">草案</td>
+				</c:if>
+				<c:if test="${draftCon.status==2}">
+					<td class="tc pointer" onclick="showDraftContract('${draftCon.id}')">正式</td>
+				</c:if>
 			</tr>
 		</c:forEach>
 	</table>
     </div>
    <div id="pagediv" align="right"></div>
-   <%--<form id="contractForm" action="${pageContext.request.contextPath}/purchaseContract/updateDraftById.html" method="post" enctype="multipart/form-data">
-   <input type="hidden" value="" id="ids" name="id"/>
-   <input type="hidden" value="2" name="status"/>
-   	<ul class="list-unstyled list-flow dnone mt10" id="numberWin">
-  		    <li class="col-md-12 ml15">
-			   <span class="span3 fl mt5"><div class="red star_red">*</div>合同批准文号：</span>
-			   <input type="text" id="apN" name="approvalNumber" value="" class="mb0 w220"/>
+   </div>
+   <form action="${pageContext.request.contextPath}/purchaseContract/toRoughContract.html" id="form2">
+    <input type="hidden" name="id" id="ids" value=""/>
+    <input type="hidden" name="status" value="1"/>
+    <ul class="list-unstyled mt10 dnone" id="numberWin">
+  		    <li class="col-md-6 col-sm-12 col-xs-12 pl15">
+			   <span class="col-md-12 col-sm-12 col-xs-12"><div class="red star_red">*</div>草稿合同上报时间：</span>
+			   <div class="input-append input_group col-sm-12 col-xs-12 p0 col-md-12">
+			     <input type="text" name="draftGitAt" id="draftGitAt" onfocus="WdatePicker({dateFmt:'yyyy-MM-dd HH:mm:ss'})" class="Wdate mb0 w220"/>
+			     <div id='gitTime' class="cue"></div>
+			   </div>
 			</li>
-			<li class="col-md-12">
-			   <span class="span3 fl mt5"><div class="red star_red">*</div>正式合同上报时间：</span>
-			   <input type="text" name="formalGitAt" id="formalGitAt" onfocus="WdatePicker({dateFmt:'yyyy-MM-dd HH:mm:ss'})" class="Wdate mb0 w220"/>
+			<li class="col-md-6 col-sm-12 col-xs-12">
+			   <span class="col-md-12 col-sm-12 col-xs-12"><div class="red star_red">*</div>草稿合同批复时间：</span>
+			   <div class="input-append input_group col-sm-12 col-xs-12 p0 col-md-12">
+			     <input type="text" name="draftReviewedAt" id="draftReviewedAt" onfocus="WdatePicker({dateFmt:'yyyy-MM-dd HH:mm:ss'})" class="Wdate mb0 w220"/>
+			     <div id='reviewTime' class="cue"></div>
+			   </div>
 			</li>
-			<li class="col-md-12">
-			   <span class="span3 fl mt5"><div class="red star_red">*</div>正式合同批复时间：</span>
-			   <input type="text" name="formalReviewedAt" id="formalReviewedAt" onfocus="WdatePicker({dateFmt:'yyyy-MM-dd HH:mm:ss'})" class="Wdate mb0 w220"/>
-			</li>
-			<li class="col-md-12 mt10">
-			   <span class="span3 fl"><div class="red star_red">*</div>上传附件：</span>
-			    <up:upload id="post_attach_up" businessId="${attachuuid}" sysKey="${attachsysKey}" typeId="${attachtypeId}" auto="true" />
-				<up:show showId="post_attach_show" businessId="${attachuuid}" sysKey="${attachsysKey}" typeId="${attachtypeId}"/>
-            </li>
-			<li class="tc col-md-12 mt20">
+			<li class="tc col-md-12 col-sm-12 col-xs-12 mt20">
 			 <input type="button" class="btn" onclick="save()" value="生成"/>
 			 <input type="button" class="btn" onclick="cancel()" value="取消"/>
 			</li>
-	 </ul>
+	</ul>
 	</form>
-	<div class="col-md-12 tc">
-	<div id="edi" class="w70p mt5 dnone tc" style="margin:0 auto">
-		 <script id="editor" name="content" type="text/plain" class=""></script>
-    </div>
-    </div>
-	<div id="pagediv" align="right"></div>
-	--%>
-   </div>
 </body>
 </html>
